+++ viprs/__init__.py
@@ -5,5 +5,5 @@ from .model.gridsearch.VIPRSGridSearch import VIPRSGridSearch
 from .model.gridsearch.HyperparameterGrid import HyperparameterGrid
 from .utils.data_utils import *
 
-__version__ = '0.1.1'
-__release_date__ = 'April 2024'
+__version__ = '0.1.2'
+__release_date__ = 'May 2024'